+++ errorlog/handler_test.go
@@ -1,4 +1,4 @@
-package log_test
+package errorlog_test
 
 import (
 	"io"
@@ -9,8 +9,8 @@ import (
 
 	"github.com/go-kit/kit/log"
 	"github.com/goph/emperror"
+	. "github.com/goph/emperror/errorlog"
 	"github.com/goph/emperror/internal"
-	. "github.com/goph/emperror/log"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -23,7 +23,7 @@ func TestHandler_Handle(t *testing.T) {
 
 	handler.Handle(err)
 
-	assert.Equal(t, "level=error msg=\"internal error\"\n", buf.String())
+	assert.Equal(t, "msg=\"internal error\"\n", buf.String())
 }
 
 func TestHandler_Handle_Context(t *testing.T) {
@@ -41,7 +41,7 @@ func TestHandler_Handle_Context(t *testing.T) {
 
 	handler.Handle(err)
 
-	assert.Equal(t, "level=error a=123 previous=\"previous error\" msg=\"internal error\"\n", buf.String())
+	assert.Equal(t, "a=123 previous=\"previous error\" msg=\"internal error\"\n", buf.String())
 }
 
 func TestHandler_Handle_MultiError(t *testing.T) {
@@ -51,11 +51,11 @@ func TestHandler_Handle_MultiError(t *testing.T) {
 	}{
 		"logfmt": {
 			log.NewLogfmtLogger,
-			"level=error msg=\"internal error\" parent=\"Multiple errors happened\"\nlevel=error msg=\"something else\" parent=\"Multiple errors happened\"\n",
+			"msg=\"internal error\" parent=\"Multiple errors happened\"\nmsg=\"something else\" parent=\"Multiple errors happened\"\n",
 		},
 		"json": {
 			log.NewJSONLogger,
-			"{\"level\":\"error\",\"msg\":\"internal error\",\"parent\":\"Multiple errors happened\"}\n{\"level\":\"error\",\"msg\":\"something else\",\"parent\":\"Multiple errors happened\"}\n",
+			"{\"msg\":\"internal error\",\"parent\":\"Multiple errors happened\"}\n{\"msg\":\"something else\",\"parent\":\"Multiple errors happened\"}\n",
 		},
 	}
 
@@ -74,4 +74,16 @@ func TestHandler_Handle_MultiError(t *testing.T) {
 			assert.Equal(t, test.expected, buf.String())
 		})
 	}
+}
+
+func TestMessageField(t *testing.T) {
+	buf := &bytes.Buffer{}
+	logger := log.NewLogfmtLogger(buf)
+	handler := NewHandler(logger, MessageField("message"))
+
+	err := errors.New("internal error")
+
+	handler.Handle(err)
+
+	assert.Equal(t, "message=\"internal error\"\n", buf.String())
 }